ibc: replace gov dep with DI
