+++ crates/ibc/Cargo.toml
@@ -24,7 +24,6 @@ testing = ["namada_core/testing", "ibc-testkit", "proptest"]
 namada_core = { path = "../core" }
 namada_events = { path = "../events", default-features = false }
 namada_gas = { path = "../gas" }
-namada_governance = { path = "../governance" }
 namada_macros = {path = "../macros"}
 namada_migrations = {path = "../migrations", optional = true}
 namada_state = { path = "../state" }
@@ -55,6 +54,7 @@ tracing.workspace = true
 
 [dev-dependencies]
 namada_core = { path = "../core", features = ["testing"] }
+namada_governance = { path = "../governance" }
 namada_parameters = { path = "../parameters", features = ["testing"] }
 namada_proof_of_stake = { path = "../proof_of_stake", features = ["testing"] }
 namada_state = { path = "../state", features = ["testing"] }
